Fix Wave Bug

+++ Assets/Scripts/UI/Damage PopUp/DamagePopUpGenerator.cs
@@ -54,7 +54,7 @@ public class DamagePopUpGenerator : Singleton<DamagePopUpGenerator>
 
 		var tmp = popup.GetComponentInChildren<TMP_Text>();
 		popup.transform.position = position;
-		tmp.text = ((int)damageInfo.Damage).ToString();
+		tmp.text = damageInfo.Damage.ToString();
 		popup.ResetPopUp();
 	}
 }
